+++ webknossos/tests/dataset/test_dataset_add_remote_mag_and_layer.py
@@ -1,4 +1,5 @@
 import itertools
+import sys
 from collections.abc import Iterable, Iterator
 from pathlib import Path
 
@@ -10,7 +11,10 @@ from webknossos.dataset import COLOR_CATEGORY, Dataset
 from webknossos.geometry import BoundingBox
 from webknossos.utils import is_remote_path
 
-pytestmark = [pytest.mark.use_proxay]
+pytestmark = [
+    pytest.mark.skipif(sys.platform == "win32", reason="too slow on windows"),
+    pytest.mark.use_proxay,
+]
 
 
 @pytest.fixture