+++ webknossos/Changelog.md
@@ -20,7 +20,7 @@ For upgrade instructions, please check the respective _Breaking Changes_ section
 - Warnings regarding missing pims image readers are only printed once during conversion. Error messages regarding wk-libs version are not displayed in slurm environments. [#1266](https://github.com/scalableminds/webknossos-libs/pull/1266)
 
 ### Fixed
-
+- Fixed mag and attachment paths under Windows. [#1329](https://github.com/scalableminds/webknossos-libs/pull/1329)
 
 ## [2.4.2](https://github.com/scalableminds/webknossos-libs/releases/tag/v2.4.2) - 2025-07-07
 [Commits](https://github.com/scalableminds/webknossos-libs/compare/v2.4.1...v2.4.2)